move api to own directory and separate endpoints
